Progress on Redirect ACK
wrote some logic for redirect messge but it isn't working yet. also sending peerPort and peerIP in PTP packet not working, the destination peer does not get the correct values, i suspect it is because the ports have their own copies of the peertables. maybe i should try storing them in the singleton