Fixed crashing on closing peers
This problem was only an issue on windows which was very strange, but regardless it can't hurt to have an error handler in the code. I simply ignore the error if it is of type ECONNRESET, which was the error type being thrown when I closed a peer on windows.